+++ karma.conf.js
@@ -1,58 +1,25 @@
-// Karma configuration
+module.exports = function (config) {
+  'use strict';
 
-// base path, that will be used to resolve files and exclude
-basePath = '';
-
-// list of files / patterns to load in the browser
-files = [
-  JASMINE,
-  JASMINE_ADAPTER,
-  'bower_components/jquery/jquery.js',
-  'bower_components/sinonjs/sinon.js',
-  'bower_components/jasmine-sinon/lib/jasmine-sinon.js',
-  'bower_components/angular/angular.js',
-  'bower_components/angular-mocks/angular-mocks.js',
-  'src/angular-masonry.js',
-  'test/mocks/**/*.js',
-  'test/spec/**/*.coffee'
-];
-
-// list of files to exclude
-exclude = [];
-
-// test results reporter to use
-// possible values: dots || progress || growl
-reporters = ['progress'];
-
-// web server port
-port = 8080;
-
-// cli runner port
-runnerPort = 9100;
-
-// enable / disable colors in the output (reporters and logs)
-colors = true;
-
-// level of logging
-// possible values: LOG_DISABLE || LOG_ERROR || LOG_WARN || LOG_INFO || LOG_DEBUG
-logLevel = LOG_INFO;
-
-// enable / disable watching file and executing tests whenever any file changes
-autoWatch = false;
-
-// Start these browsers, currently available:
-// - Chrome
-// - ChromeCanary
-// - Firefox
-// - Opera
-// - Safari (only Mac)
-// - PhantomJS
-// - IE (only Windows)
-browsers = ['Chrome'];
-
-// If browser does not capture in given timeout [ms], kill it
-captureTimeout = 5000;
-
-// Continuous Integration mode
-// if true, it capture browsers, run tests and exit
-singleRun = true;
+  config.set({
+    basePath: '',
+    frameworks: ['jasmine'],
+    files: [
+      'bower_components/jquery/jquery.js',
+      'bower_components/sinonjs/sinon.js',
+      'bower_components/jasmine-sinon/lib/jasmine-sinon.js',
+      'bower_components/angular/angular.js',
+      'bower_components/angular-mocks/angular-mocks.js',
+      'src/angular-masonry.js',
+      'test/mocks/**/*.js',
+      'test/spec/**/*.coffee'
+    ],
+    exclude: [],
+    reporters: ['dots'],
+    autoWatch: false,
+    browsers: ['Chrome'],
+    captureTimeout: 5000,
+    singleRun: true,
+    reportSlowerThan: 100
+  });
+};