+++ src/angular-masonry.js
@@ -16,7 +16,7 @@
 
       this.scheduleMasonryOnce = function scheduleMasonryOnce() {
         var args = arguments;
-        var found = schedule.filter(function (item) {
+        var found = schedule.filter(function filterFn(item) {
           return item[0] === args[0];
         }).length > 0;
 
@@ -38,7 +38,7 @@
           if (destroyed) {
             return;
           }
-          schedule.forEach(function (args) {
+          schedule.forEach(function scheduleForEach(args) {
             $element.masonry.apply($element, args);
           });
           schedule = [];
@@ -69,6 +69,7 @@
             // Keep track of added elements.
             bricks[id] = true;
             $element.masonry('appended', element, true);
+            self.scheduleMasonryOnce('reloadItems');
             self.scheduleMasonryOnce('layout');
           }
         }
@@ -98,7 +99,13 @@
         bricks = [];
       };
 
-    }).directive('masonry', function () {
+      this.reload = function reload() {
+        $element.masonry();
+        $scope.$emit('masonry.reloaded');
+      };
+
+
+    }).directive('masonry', function masonryDirective() {
       return {
         restrict: 'AE',
         controller: 'MasonryCtrl',
@@ -116,18 +123,30 @@
           }
         }
       };
-    }).directive('masonryBrick', function () {
+    }).directive('masonryBrick', function masonryBrickDirective() {
       return {
         restrict: 'AC',
         require: '^masonry',
         scope: true,
         link: {
           pre: function preLink(scope, element, attrs, ctrl) {
-            var id = scope.$id;
+            var id = scope.$id, index;
 
             ctrl.appendBrick(element, id);
             element.on('$destroy', function () {
               ctrl.removeBrick(id, element);
+            });
+            
+            scope.$on('masonry.reload', function () {
+              ctrl.reload();
+            });
+
+            scope.$watch('$index', function () {
+              if (index !== undefined && index !== scope.$index) {
+                ctrl.scheduleMasonryOnce('reloadItems');
+                ctrl.scheduleMasonryOnce('layout');
+              }
+              index = scope.$index;
             });
           }
         }